PropertyEncoder#encodeLastCoConstraints()
Move RelCo#coClockVar(Event,SolverContext) into Utils
