+++ dartagnan/src/main/java/com/dat3m/dartagnan/encoding/PropertyEncoder.java
@@ -1,11 +1,14 @@
 package com.dat3m.dartagnan.encoding;
 
 import com.dat3m.dartagnan.configuration.Property;
+import com.dat3m.dartagnan.expression.IExpr;
 import com.dat3m.dartagnan.program.Program;
 import com.dat3m.dartagnan.program.Thread;
+import com.dat3m.dartagnan.program.analysis.ExecutionAnalysis;
 import com.dat3m.dartagnan.program.analysis.alias.AliasAnalysis;
 import com.dat3m.dartagnan.program.event.Tag;
 import com.dat3m.dartagnan.program.event.core.Event;
+import com.dat3m.dartagnan.program.event.core.Init;
 import com.dat3m.dartagnan.program.event.core.Label;
 import com.dat3m.dartagnan.program.event.core.Load;
 import com.dat3m.dartagnan.program.event.core.MemEvent;
@@ -14,10 +17,11 @@ import com.dat3m.dartagnan.program.filter.FilterMinus;
 import com.dat3m.dartagnan.verification.Context;
 import com.dat3m.dartagnan.wmm.Wmm;
 import com.dat3m.dartagnan.wmm.axiom.Axiom;
+import com.dat3m.dartagnan.wmm.relation.Relation;
 import com.dat3m.dartagnan.wmm.relation.RelationNameRepository;
-import com.dat3m.dartagnan.wmm.relation.base.memory.RelCo;
 import com.dat3m.dartagnan.wmm.relation.base.memory.RelRf;
 import com.dat3m.dartagnan.wmm.utils.Tuple;
+import com.dat3m.dartagnan.wmm.utils.TupleSet;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.sosy_lab.common.configuration.Configuration;
@@ -25,6 +29,7 @@ import org.sosy_lab.common.configuration.InvalidConfigurationException;
 import org.sosy_lab.common.configuration.Options;
 import org.sosy_lab.java_smt.api.BooleanFormula;
 import org.sosy_lab.java_smt.api.BooleanFormulaManager;
+import org.sosy_lab.java_smt.api.Formula;
 import org.sosy_lab.java_smt.api.IntegerFormulaManager;
 import org.sosy_lab.java_smt.api.SolverContext;
 
@@ -34,6 +39,11 @@ import java.util.stream.Collectors;
 
 import static com.dat3m.dartagnan.configuration.Property.*;
 import static com.dat3m.dartagnan.expression.utils.Utils.generalEqual;
+import static com.dat3m.dartagnan.program.Program.SourceLanguage.LITMUS;
+import static com.dat3m.dartagnan.program.event.Tag.INIT;
+import static com.dat3m.dartagnan.program.event.Tag.WRITE;
+import static com.dat3m.dartagnan.wmm.analysis.RelationAnalysis.findTransitivelyImpliedCo;
+import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.CO;
 import static com.dat3m.dartagnan.wmm.utils.Utils.edge;
 import static com.dat3m.dartagnan.wmm.utils.Utils.intVar;
 import static com.google.common.base.Preconditions.*;
@@ -46,6 +56,7 @@ public class PropertyEncoder implements Encoder {
 
     private final Program program;
     private final Wmm memoryModel;
+    private final ExecutionAnalysis exec;
     private final AliasAnalysis alias;
 
     // =====================================================================
@@ -55,6 +66,7 @@ public class PropertyEncoder implements Encoder {
                 "The program must get compiled first before its properties can be encoded.");
         this.program = checkNotNull(program);
         this.memoryModel = checkNotNull(wmm);
+        this.exec = context.requires(ExecutionAnalysis.class);
         this.alias = context.requires(AliasAnalysis.class);
         config.inject(this);
     }
@@ -82,6 +94,9 @@ public class PropertyEncoder implements Encoder {
     	if(property.contains(CAT)) {
     		enc = bmgr.or(enc, encodeCATProperties(ctx));
     	}
+        if (program.getFormat().equals(LITMUS) || property.contains(LIVENESS)) {
+            enc = bmgr.and(enc, encodeLastCoConstraints(ctx));
+        }
     	return enc;
     }
 
@@ -165,7 +180,6 @@ public class PropertyEncoder implements Encoder {
 
         BooleanFormulaManager bmgr = ctx.getFormulaManager().getBooleanFormulaManager();
         RelRf rf = (RelRf) memoryModel.getRelation(RelationNameRepository.RF);
-        RelCo co = (RelCo) memoryModel.getRelation(RelationNameRepository.CO);
         // Compute "stuckness": A thread is stuck if it reaches a spinloop bound event
         // while reading from a co-maximal write.
         Map<Thread, BooleanFormula> isStuckMap = new HashMap<>();
@@ -181,7 +195,7 @@ public class PropertyEncoder implements Encoder {
                 for (Load load : pair.loads) {
                     BooleanFormula coMaximalLoad = bmgr.makeFalse();
                     for (Tuple rfEdge : rf.getMaxTupleSet().getBySecond(load)) {
-                        coMaximalLoad = bmgr.or(coMaximalLoad, bmgr.and(rf.getSMTVar(rfEdge, ctx), co.getLastCoVar(rfEdge.getFirst(), ctx)));
+                        coMaximalLoad = bmgr.or(coMaximalLoad, bmgr.and(rf.getSMTVar(rfEdge, ctx), lastCoVar(rfEdge.getFirst(), ctx)));
                     }
                     allCoMaximalLoad = bmgr.and(allCoMaximalLoad, coMaximalLoad);
                 }
@@ -247,5 +261,66 @@ public class PropertyEncoder implements Encoder {
 		enc = bmgr.equivalence(RACES.getSMTVariable(ctx), enc);
 		// No need to use the SMT variable if the formula is trivially false 
         return bmgr.isFalse(enc) ? enc : bmgr.and(RACES.getSMTVariable(ctx), enc);
+    }
+
+    private BooleanFormula encodeLastCoConstraints(SolverContext ctx) {
+        final Relation co = memoryModel.getRelation(CO);
+        final BooleanFormulaManager bmgr = ctx.getFormulaManager().getBooleanFormulaManager();
+        final TupleSet minSet = co.getMinTupleSet();
+        final TupleSet maxSet = co.getMaxTupleSet();
+        final List<Event> initEvents = program.getCache().getEvents(FilterBasic.get(INIT));
+        final List<Event> writes = program.getCache().getEvents(FilterBasic.get(WRITE));
+        final boolean doEncodeFinalAddressValues = program.getFormat() == LITMUS;
+        // Find transitively implied coherences. We can use these to reduce the encoding.
+        final Set<Tuple> transCo = findTransitivelyImpliedCo(co, exec);
+        // Find all writes that are never last, i.e., those that will always have a co-successor.
+        final Set<Event> dominatedWrites = minSet.stream()
+                .filter(t -> exec.isImplied(t.getFirst(), t.getSecond()))
+                .map(Tuple::getFirst).collect(Collectors.toSet());
+        // ---- Construct encoding ----
+        BooleanFormula enc = bmgr.makeTrue();
+        for (Event writeEvent : writes) {
+            MemEvent w1 = (MemEvent) writeEvent;
+            if (dominatedWrites.contains(w1)) {
+                enc = bmgr.and(enc, bmgr.equivalence(lastCoVar(w1, ctx), bmgr.makeFalse()));
+                continue;
+            }
+            BooleanFormula isLast = w1.exec();
+            // ---- Find all possibly overwriting writes ----
+            for (Tuple t : maxSet.getByFirst(w1)) {
+                if (transCo.contains(t)) {
+                    // We can skip the co-edge (w1,w2), because there will be an intermediate write w3
+                    // that already witnesses that w1 is not last.
+                    continue;
+                }
+                Event w2 = t.getSecond();
+                BooleanFormula isAfter = minSet.contains(t) ? bmgr.not(w2.exec()) : bmgr.not(co.getSMTVar(t, ctx));
+                isLast = bmgr.and(isLast, isAfter);
+            }
+            BooleanFormula lastCoExpr = lastCoVar(w1, ctx);
+            enc = bmgr.and(enc, bmgr.equivalence(lastCoExpr, isLast));
+            if (doEncodeFinalAddressValues) {
+                // ---- Encode final values of addresses ----
+                for (Event initEvent : initEvents) {
+                    Init init = (Init) initEvent;
+                    if (!alias.mayAlias(w1, init)) {
+                        continue;
+                    }
+                    IExpr address = init.getAddress();
+                    Formula a1 = w1.getMemAddressExpr();
+                    Formula a2 = address.toIntFormula(init, ctx);
+                    BooleanFormula sameAddress = alias.mustAlias(init, w1) ? bmgr.makeTrue() : generalEqual(a1, a2, ctx);
+                    Formula v1 = w1.getMemValueExpr();
+                    Formula v2 = init.getBase().getLastMemValueExpr(ctx, init.getOffset());
+                    BooleanFormula sameValue = generalEqual(v1, v2, ctx);
+                    enc = bmgr.and(enc, bmgr.implication(bmgr.and(lastCoExpr, sameAddress), sameValue));
+                }
+            }
+        }
+        return enc;
+    }
+
+    private BooleanFormula lastCoVar(Event write, SolverContext ctx) {
+        return ctx.getFormulaManager().getBooleanFormulaManager().makeVariable("co_last(" + write.repr() + ")");
     }
 }
