+++ dartagnan/src/main/java/com/dat3m/dartagnan/wmm/relation/base/memory/RelCo.java
@@ -1,11 +1,9 @@
 package com.dat3m.dartagnan.wmm.relation.base.memory;
 
-import com.dat3m.dartagnan.expression.IExpr;
 import com.dat3m.dartagnan.program.analysis.ExecutionAnalysis;
 import com.dat3m.dartagnan.program.analysis.alias.AliasAnalysis;
 import com.dat3m.dartagnan.program.event.EventCache;
 import com.dat3m.dartagnan.program.event.core.Event;
-import com.dat3m.dartagnan.program.event.core.Init;
 import com.dat3m.dartagnan.program.event.core.MemEvent;
 import com.dat3m.dartagnan.program.filter.FilterBasic;
 import com.dat3m.dartagnan.program.filter.FilterMinus;
@@ -13,8 +11,6 @@ import com.dat3m.dartagnan.wmm.analysis.WmmAnalysis;
 import com.dat3m.dartagnan.wmm.relation.Relation;
 import com.dat3m.dartagnan.wmm.utils.Tuple;
 import com.dat3m.dartagnan.wmm.utils.TupleSet;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Predicate;
 import com.google.common.collect.Lists;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -26,17 +22,14 @@ import org.sosy_lab.java_smt.api.*;
 import org.sosy_lab.java_smt.api.NumeralFormula.IntegerFormula;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 import static com.dat3m.dartagnan.configuration.OptionNames.IDL_TO_SAT;
-import static com.dat3m.dartagnan.configuration.Property.LIVENESS;
 import static com.dat3m.dartagnan.expression.utils.Utils.generalEqual;
-import static com.dat3m.dartagnan.program.Program.SourceLanguage.LITMUS;
 import static com.dat3m.dartagnan.program.event.Tag.INIT;
 import static com.dat3m.dartagnan.program.event.Tag.WRITE;
+import static com.dat3m.dartagnan.wmm.analysis.RelationAnalysis.findTransitivelyImpliedCo;
 import static com.dat3m.dartagnan.wmm.relation.RelationNameRepository.CO;
-import static com.dat3m.dartagnan.wmm.utils.Utils.intVar;
-import static org.sosy_lab.java_smt.api.FormulaType.BooleanType;
+import static com.dat3m.dartagnan.wmm.utils.Utils.*;
 
 @Options
 public class RelCo extends Relation {
@@ -126,15 +119,7 @@ public class RelCo extends Relation {
 
     @Override
     protected BooleanFormula encodeApprox(SolverContext ctx) {
-        final BooleanFormulaManager bmgr = ctx.getFormulaManager().getBooleanFormulaManager();
-        final boolean doEncodeLastCo = task.getProgram().getFormat().equals(LITMUS) || task.getProperty().contains(LIVENESS);
-
-        BooleanFormula enc = useSATEncoding ? encodeSAT(ctx) : encodeIDL(ctx);
-        if (doEncodeLastCo) {
-            enc = bmgr.and(enc, encodeLastCoConstraints(ctx));
-        }
-
-        return enc;
+        return useSATEncoding ? encodeSAT(ctx) : encodeIDL(ctx);
     }
 
     private BooleanFormula encodeIDL(SolverContext ctx) {
@@ -145,13 +130,13 @@ public class RelCo extends Relation {
         final List<MemEvent> allWrites = new ArrayList<>(Lists.transform(cache.getEvents(FilterBasic.get(WRITE)), MemEvent.class::cast));
         allWrites.sort(Comparator.comparingInt(Event::getCId));
         final TupleSet maxSet = getMaxTupleSet();
-        final Set<Tuple> transCo = findTransitivelyImpliedCo();
+        final Set<Tuple> transCo = findTransitivelyImpliedCo(this, exec);
 
         BooleanFormula enc = bmgr.makeTrue();
         // ---- Encode clock conditions (init = 0, non-init > 0) ----
         IntegerFormula zero = imgr.makeNumber(0);
         for (MemEvent w : allWrites) {
-            IntegerFormula clock = getClockVar(w, ctx);
+            IntegerFormula clock = coClockVar(w, ctx);
             enc = bmgr.and(enc, w.is(INIT) ? imgr.equal(clock, zero) : imgr.greaterThan(clock, zero));
         }
 
@@ -171,9 +156,9 @@ public class RelCo extends Relation {
                         generalEqual(w1.getMemAddressExpr(), w2.getMemAddressExpr(), ctx);
                 BooleanFormula pairingCond = bmgr.and(execPair, sameAddress);
                 BooleanFormula fCond = (w1.is(INIT) || transCo.contains(t)) ? bmgr.makeTrue() :
-                        imgr.lessThan(getClockVar(w1, ctx), getClockVar(w2, ctx));
+                        imgr.lessThan(coClockVar(w1, ctx), coClockVar(w2, ctx));
                 BooleanFormula bCond = (w2.is(INIT) || transCo.contains(t.getInverse())) ? bmgr.makeTrue() :
-                        imgr.lessThan(getClockVar(w2, ctx), getClockVar(w1, ctx));
+                        imgr.lessThan(coClockVar(w2, ctx), coClockVar(w1, ctx));
                 BooleanFormula coF = forwardPossible ? getSMTVar(w1, w2, ctx) : bmgr.makeFalse();
                 BooleanFormula coB = backwardPossible ? getSMTVar(w2, w1, ctx) : bmgr.makeFalse();
 
@@ -242,119 +227,5 @@ public class RelCo extends Relation {
         }
 
         return enc;
-    }
-
-    private BooleanFormula encodeLastCoConstraints(SolverContext ctx) {
-        final AliasAnalysis alias = analysisContext.requires(AliasAnalysis.class);
-        final ExecutionAnalysis exec = analysisContext.requires(ExecutionAnalysis.class);
-        final BooleanFormulaManager bmgr = ctx.getFormulaManager().getBooleanFormulaManager();
-        final TupleSet minSet = getMinTupleSet();
-        final TupleSet maxSet = getMaxTupleSet();
-        final EventCache cache = task.getProgram().getCache();
-        final List<Init> initEvents = Lists.transform(cache.getEvents(FilterBasic.get(INIT)), Init.class::cast);
-        final List<MemEvent> writes = Lists.transform(cache.getEvents(FilterBasic.get(WRITE)), MemEvent.class::cast);
-        final boolean doEncodeFinalAddressValues = task.getProgram().getFormat() == LITMUS;
-
-        // Find transitively implied coherences. We can use these to reduce the encoding.
-        final Set<Tuple> transCo = findTransitivelyImpliedCo();
-        // Find all writes that are never last, i.e., those that will always have a co-successor.
-        final Set<Event> dominatedWrites = minSet.stream()
-                .filter(t -> exec.isImplied(t.getFirst(), t.getSecond()))
-                .map(Tuple::getFirst).collect(Collectors.toSet());
-
-
-        // ---- Construct encoding ----
-        BooleanFormula enc = bmgr.makeTrue();
-        for (MemEvent w1 : writes) {
-            if (dominatedWrites.contains(w1)) {
-                enc = bmgr.and(enc, bmgr.equivalence(getLastCoVar(w1, ctx), bmgr.makeFalse()));
-                continue;
-            }
-
-            BooleanFormula isLast = w1.exec();
-            // ---- Find all possibly overwriting writes ----
-            for (Tuple t : maxSet.getByFirst(w1)) {
-                if (transCo.contains(t)) {
-                    // We can skip the co-edge (w1,w2), because there will be an intermediate write w3
-                    // that already witnesses that w1 is not last.
-                    continue;
-                }
-                Event w2 = t.getSecond();
-                BooleanFormula isAfter = minSet.contains(t) ? bmgr.not(w2.exec()) : bmgr.not(getSMTVar(t, ctx));
-                isLast = bmgr.and(isLast, isAfter);
-            }
-            BooleanFormula lastCoExpr = getLastCoVar(w1, ctx);
-            enc = bmgr.and(enc, bmgr.equivalence(lastCoExpr, isLast));
-
-            if (doEncodeFinalAddressValues) {
-                // ---- Encode final values of addresses ----
-                for (Init init : initEvents) {
-                    if (!alias.mayAlias(w1, init)) {
-                        continue;
-                    }
-                    IExpr address = init.getAddress();
-                    Formula a1 = w1.getMemAddressExpr();
-                    Formula a2 = address.toIntFormula(init, ctx);
-                    BooleanFormula sameAddress = alias.mustAlias(init, w1) ? bmgr.makeTrue() : generalEqual(a1, a2, ctx);
-                    Formula v1 = w1.getMemValueExpr();
-                    Formula v2 = init.getBase().getLastMemValueExpr(ctx, init.getOffset());
-                    BooleanFormula sameValue = generalEqual(v1, v2, ctx);
-                    enc = bmgr.and(enc, bmgr.implication(bmgr.and(lastCoExpr, sameAddress), sameValue));
-                }
-            }
-        }
-        return enc;
-    }
-
-    /*
-        Returns a set of co-edges (w1, w2) (subset of maxTupleSet) whose clock-constraints
-        do not need to get encoded explicitly.
-        The reason is that whenever we have co(w1,w2) then there exists an intermediary
-        w3 s.t. co(w1, w3) /\ co(w3, w2). As a result we have c(w1) < c(w3) < c(w2) transitively.
-        Reasoning: Let (w1, w2) be a potential co-edge. Suppose there exists a w3 different to w1 and w2,
-        whose execution is either implied by either w1 or w2.
-        Now, if co(w1, w3) is a must-edge and co(w2, w3) is impossible, then we can reason as follows.
-            - Suppose w1 and w2 get executed and their addresses match, then w3 must also get executed.
-            - Since co(w1, w3) is a must-edge, we have that w3 accesses the same address as w1 and w2,
-              and c(w1) < c(w3).
-            - Because addr(w2)==addr(w3), we must also have either co(w2, e3) or co(w3, w2).
-              The former is disallowed by assumption, so we have co(w3, w2) and hence c(w3) < c(w2).
-            - By transitivity, we have c(w1) < c(w3) < c(w2) as desired.
-            - Note that this reasoning has to be done inductively, because co(w1, w3) or co(w3, w2) may
-              not involve encoding a clock constraint (due to this optimization).
-        There is also a symmetric case where co(w3, w1) is impossible and co(w3, w2) is a must-edge.
-
-     */
-    private Set<Tuple> findTransitivelyImpliedCo() {
-        final ExecutionAnalysis exec = analysisContext.requires(ExecutionAnalysis.class);
-        final TupleSet min = getMinTupleSet();
-        final TupleSet max = getMaxTupleSet();
-
-        Set<Tuple> transCo = new HashSet<>();
-        for (final Tuple t : max) {
-            final MemEvent e1 = (MemEvent) t.getFirst();
-            final MemEvent e2 = (MemEvent) t.getSecond();
-            final Predicate<Event> execPred = (e3 -> e3 != e1 && e3 != e2 && (exec.isImplied(e1, e3) || exec.isImplied(e2, e3)));
-            final boolean hasIntermediary = min.getByFirst(e1).stream().map(tuple -> (MemEvent)tuple.getSecond())
-                                    .anyMatch(e3 -> execPred.apply(e3) && !max.contains(new Tuple(e2, e3))) ||
-                                min.getBySecond(e2).stream().map(tuple -> (MemEvent)tuple.getFirst())
-                                    .anyMatch(e3 -> execPred.apply(e3) && !max.contains(new Tuple(e3, e1)));
-            if (hasIntermediary) {
-                transCo.add(t);
-            }
-        }
-        return transCo;
-    }
-
-    public IntegerFormula getClockVar(Event write, SolverContext ctx) {
-    	Preconditions.checkArgument(write.is(WRITE), "Cannot get a clock-var for non-writes.");
-        if (write.is(INIT)) {
-            return ctx.getFormulaManager().getIntegerFormulaManager().makeNumber(0);
-        }
-        return intVar(term, write, ctx);
-    }
-
-    public BooleanFormula getLastCoVar(Event write, SolverContext ctx) {
-        return ctx.getFormulaManager().makeVariable(BooleanType, "co_last(" + write.repr() + ")");
     }
 }
